Add granular permissions to Status -> Services.
git-svn-id: http://fusionpbx.googlecode.com/svn/trunk/fusionpbx@1143 e2954ef0-caed-11de-bd98-e341d93cc3cc

+++ mod/services/v_services.php
@@ -26,7 +26,7 @@
 require_once "root.php";
 require_once "includes/config.php";
 require_once "includes/checkauth.php";
-if (ifgroup("superadmin")) {
+if (permission_exists('services_view')) {
 	//access granted
 }
 else {
@@ -95,7 +95,6 @@ if (strlen($_GET["a"]) > 0) {
 	echo "	<td align=\"center\">\n";
 	echo "		<br>";
 
-
 	echo "<table width='100%' border='0'>\n";
 	echo "<tr>\n";
 	echo "<td width='50%' align='left' nowrap='nowrap'><b>Services</b></td>\n";
@@ -107,7 +106,6 @@ if (strlen($_GET["a"]) > 0) {
 	echo "</td>\n";
 	echo "</tr>\n";
 	echo "</tr></table>\n";
-
 
 	$sql = "";
 	$sql .= " select * from v_services ";
@@ -134,32 +132,32 @@ if (strlen($_GET["a"]) > 0) {
 	$resultcount = count($result);
 	unset ($prepstatement, $sql);
 
-
 	$c = 0;
 	$rowstyle["0"] = "rowstyle0";
 	$rowstyle["1"] = "rowstyle1";
 
 	echo "<div align='center'>\n";
 	echo "<table width='100%' border='0' cellpadding='0' cellspacing='0'>\n";
-
 	echo "<tr>\n";
 	echo thorderby('v_service_name', 'Name', $orderby, $order);
 	echo thorderby('v_service_desc', 'Description', $orderby, $order);
 	echo "<th>Status</th>\n";
 	echo "<th>Action</th>\n";
 	echo "<td align='right' width='42'>\n";
-	echo "	<a href='v_services_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('services_add')) {
+		echo "	<a href='v_services_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "</td>\n";
 	echo "<tr>\n";
 
-	if ($resultcount == 0) { //no results
+	if ($resultcount == 0) {
+		//no results
 	}
 	else { //received results
 		foreach($result as $row) {
 			echo "<tr >\n";
 			echo "	<td valign='top' class='".$rowstyle[$c]."'>".$row[v_service_name]."</td>\n";
 			echo "	<td valign='top' class='".$rowstyle[$c]."'>".$row[v_service_desc]."</td>\n";
-
 			echo "	<td valign='top' class='".$rowstyle[$c]."'>\n";
 			$pid = file_get_contents($row[v_service_data]);
 			if (is_process_running($pid)) {
@@ -169,7 +167,6 @@ if (strlen($_GET["a"]) > 0) {
 				echo "<strong>Stopped</strong>";
 			}
 			echo "</td>\n";
-
 			echo "	<td valign='top' class='".$rowstyle[$c]."'>\n";
 			if (is_process_running($pid)) {
 				echo "		<a href='v_services.php?id=".$row[service_id]."&a=stop' alt='stop'>Stop</a>";
@@ -178,17 +175,19 @@ if (strlen($_GET["a"]) > 0) {
 				echo "		<a href='v_services.php?id=".$row[service_id]."&a=start' alt='start'>Start</a>";
 			}
 			echo "</td>\n";
-
 			echo "	<td valign='top' align='right'>\n";
-			echo "		<a href='v_services_edit.php?id=".$row[service_id]."' alt='edit'>$v_link_label_edit</a>\n";
-			echo "		<a href='v_services_delete.php?id=".$row[service_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
+			if (permission_exists('services_edit')) {
+				echo "		<a href='v_services_edit.php?id=".$row[service_id]."' alt='edit'>$v_link_label_edit</a>\n";
+			}
+			if (permission_exists('services_delete')) {
+				echo "		<a href='v_services_delete.php?id=".$row[service_id]."' alt='delete' onclick=\"return confirm('Do you really want to delete this?')\">$v_link_label_delete</a>\n";
+			}
 			echo "	</td>\n";
 			echo "</tr>\n";
 			if ($c==0) { $c=1; } else { $c=0; }
 		} //end foreach
 		unset($sql, $result, $rowcount);
 	} //end if results
-
 
 	echo "<tr>\n";
 	echo "<td colspan='5' align='left'>\n";
@@ -197,19 +196,19 @@ if (strlen($_GET["a"]) > 0) {
 	echo "		<td width='33.3%' nowrap>&nbsp;</td>\n";
 	echo "		<td width='33.3%' align='center' nowrap>$pagingcontrols</td>\n";
 	echo "		<td width='33.3%' align='right'>\n";
-	echo "			<a href='v_services_edit.php' alt='add'>$v_link_label_add</a>\n";
+	if (permission_exists('services_add')) {
+		echo "			<a href='v_services_edit.php' alt='add'>$v_link_label_add</a>\n";
+	}
 	echo "		</td>\n";
 	echo "	</tr>\n";
  	echo "	</table>\n";
 	echo "</td>\n";
 	echo "</tr>\n";
 
-
 	echo "</table>";
 	echo "</div>";
 	echo "<br><br>";
 	echo "<br><br>";
-
 
 	echo "</td>";
 	echo "</tr>";
@@ -217,11 +216,7 @@ if (strlen($_GET["a"]) > 0) {
 	echo "</div>";
 	echo "<br><br>";
 
+//include the footer
+	require_once "includes/footer.php";
 
-require_once "includes/footer.php";
-unset ($resultcount);
-unset ($result);
-unset ($key);
-unset ($val);
-unset ($c);
 ?>
